+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Changed
 - try-catch refactoring in `ReportPortal.Builder` class to get better stack traces
+### Added
+- `rp.rx.buffer.size` property handling
 
 ## [5.0.18]
 ### Added